requirements.txt
abydos==0.5.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
audioread==3.0.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.32.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
editdistance==0.8.1
eng_to_ipa==0.0.2
epitran==1.25.1
fastapi==0.115.6
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
g2pK==0.9.4
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gssapi==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.7
ipa==4.12.2
ipaclient==4.12.2
ipalib==4.12.2
ipaplatform==4.12.2
ipapython==4.12.2
isodate==0.7.2
itsdangerous==2.2.0
jamo==0.4.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jpype1==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
konlpy==0.6.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
language-tool-python==2.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.14.3
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
msrest==0.7.1
multidict==6.1.0
munkres==1.1.4
mypy-extensions==1.0.0
narwhals==1.13.5
netaddr==1.3.0
networkx==3.4.2
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.59.7
orjson==3.10.14
packaging==24.0
pandas==2.2.3
panphon==0.21.2
phonetics==1.0.5
pickle4==0.0.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-multipart==0.0.17
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.4.0
scipy==1.13.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.37
starlette==0.41.3
streamlit==1.40.1
streamlit-lottie==0.0.3
sympy==1.13.1
tenacity==9.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.7.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unicodecsv==0.14.1
urllib3==2.2.1
uvicorn==0.32.0
watchdog==6.0.0
Werkzeug==3.0.3
wheel==0.45.0
yarl==1.18.3
zones==0.0.2
